requirements.txt
asgiref==3.3.0
beautifulsoup4==4.9.3
Django==3.1.2
django-bootstrap4==2.3.1
django-octicons==1.0.2
importlib-metadata==2.0.0
mysqlclient==2.0.1
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.4.1
zipp==3.3.1
social-auth-app-django==4.0.0
randomcolor
arrow
ntplib
django-filter==2.4.0
django-model-utils